requirements.txt
lxml
pycryptodome
oscrypto
asn1crypto
